reorganizes folder structure
