--- src/About/index.js
@@ -1,2 +1,0 @@
-import About from './About'
-export default About
