--- src/Create/index.js
@@ -1,3 +1,0 @@
-import Create from './Create'
-
-export default Create
