+++ src/_Create/index.js
@@ -1,0 +1,2 @@
+import Container from './Container'
+export default Container
